Move normalization functions to text_formatter
